% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{dirichlet_energy}
\alias{dirichlet_energy}
\title{Dirichlet Energy}
\usage{
dirichlet_energy(x, standartize = FALSE)
}
\arguments{
\item{x}{data. Vecotrs}

\item{standartize}{Logical. If set to TRUE, then the final result is divided by the amount of entries. Defaults to FALSE.}
}
\value{
The signal's Dirichlet Energy. If x is a data frame, then the output is a vector.
}
\description{
Calculates the Dirichlet Energy of a data vector.
}
\examples{

a <- seq(1, 100)
dirichlet_energy(a)

}
