% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian_norm.R
\name{euclidian_norm}
\alias{euclidian_norm}
\title{Euclidian Norm}
\usage{
euclidian_norm(v)
}
\arguments{
\item{v}{must be a numerical vector}
}
\value{
The vector's norm
}
\description{
Calculates the euclidian norm of a vector
}
\examples{

v = rnorm(n = 200)
euclidian_norm(v)

}
