% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gram_Schmidt.R
\name{GS_process}
\alias{GS_process}
\title{Gram-Schmidt Process}
\usage{
GS_process(X, type = "orthogonal")
}
\arguments{
\item{X}{a matrix of numerical vectors}

\item{type}{Can be either "orthonormal" or "orthogonal". If "Orthonormal" then resulting vectores are all normalized to the standard basis.}
}
\value{
The resulting matrix
}
\description{
Applies the Gram-Schimdt Process to a set of vectors
}
\examples{

v = rnorm(n = 200)
A = matrix(v, ncol = 5)

B = GS_process(A)

}
