% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_and_pair.R
\name{is_even}
\alias{is_even}
\title{Checking if number is even}
\usage{
is_even(x)
}
\arguments{
\item{x}{vector of integers or a single integer.}
}
\value{
Logical, true if number is integer.
}
\description{
Evaluates the evenness of a number.
}
\examples{

is_even(rbinom(1, size = 2, prob = .5))

}
