% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd_and_pair.R
\name{is_odd}
\alias{is_odd}
\title{Checking if number is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{vector of integers or a single integer.}
}
\value{
Logical, true if number is odd.
}
\description{
Evaluates the oddness of a number.
}
\examples{

is_odd(rbinom(1, size = 2, prob = .5))

}
