% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weierstrass.R
\name{weierstrass}
\alias{weierstrass}
\title{Weierstrass Function}
\usage{
weierstrass(x, a = 0.5, b = 15, eps = 0.01, max = 1000)
}
\arguments{
\item{x}{value in which in function should be evaluated. Can be an integer, double or vector.}

\item{a}{parameter of the function. Must be positive and smaller than 1. Defaults to .5.}

\item{b}{parameter of the function. Must be positive and odd. Defaults to 5. For convergence it is necessary that ab be larger than (1 + (3/2)pi).}

\item{eps}{parameter for convergence of the calculation. Series is evaluated up until the marginal value is smaller than eps. Defaults to .001.}

\item{max}{parameter for convergence of the calculation. Maximum number of iterations. Defaults to 100.}
}
\value{
Output of the Weierstrass Function
}
\description{
Evaluates the Weierstrass Function
}
\examples{

a = seq(1, 100)
weierstrass(a)

}
